[RFC] init: introduce service group option "shutdown"
The init code currently maintains a list of hard-coded services that
should either:
1. not be killed during reboot at all, but started.
2. not be killed until after apps have been killed, and not started.
3. vold be killed after apps but needs to be started and get sigterm.

ISSUES STILL:
1. bootanim, see the large message TODO block.
2. Should after_apps sigkill|sigterm go to stop|terminate respectively to match the hight
   init service semantic?

I'm on vacation! Yeah... be back 4/19. Ill review all the comments and pick back up.
I Did a rebase right before uploading this, who knows what broke, no build nor test!

This poses a few issues:

1. This feature is hard coded into init.
2. Vendor services cannot leverage this.

For item 2, a specific instance in maintaining state between
the IoC and the SoC via CAN bus messages requires that the
service stays alive to send the reboot message from the SoC
to the IoC. where the SoC is the Android device.

Test: reboot the device and check pstore/console-ramoops
      to verify that reboot sequence attempted to start
      shutdown never start services like vold as well as stopping
      after_apps services like, tombstoned, logd, and adbd.
      Tested on Hikey, thus watchdogd is not available.

Change-Id: Idac46aec4f14f0fee7675c1a93e4f71869e892d1
Signed-off-by: William Roberts <william.c.roberts@intel.com>