init: introduce service group option "shutdown"
The init code currently maintains a list of hard-coded services that
should not be killed during reboot. This poses two issues:

1. This feature is hard coded into init.
   a. On devices without watchdogd, we used to get a warning:
      "Shutdown critical service not found: watchdogd"

2. Vendor services cannot leverage this.

For item 2, a specific instance in maintaining state between
the IoC and the SoC via CAN bus messages requires that the
service stays alive to send the reboot message from the SoC
to the IoC. where the SoC is the Android device.

Alternative implementation would be to do the keyword
as "shutdown_critical" vs "shutdown" that takes an
argument.

Test: reboot the device and check pstore/console-ramoops
      to verify that reboot did not kill vold. It helps
      to add an info message to the system/core/init/reboot.cpp
      file on the callback: OonShutdownCritical to ensure that
      critical services are being started and skipped.
      Tested on Hikey, thus the only service available was vold.

Change-Id: Idac46aec4f14f0fee7675c1a93e4f71869e892d1
Signed-off-by: William Roberts <william.c.roberts@intel.com>