Small bugfix and update tests.
The tls data is not marked in any special way, so add the keys as
root vals.

Add the ability to indicate where to look for leaks. This helps with
the unit tests to make sure that if something winds up on the stack
accidentally.

Update the tests to make them deterministic and pass consistently.
This almost completely works, but the 64 bit tests can still cause
some variability. I plan a follow-up cl to fix these by making the
tests run in an isolated mode.

Also, set .clang-format to the .clang-format-2 defaults.

Bug: 31919199

Test: Ran 32 bit and 64 bit unit tests on a sailfish device.
Change-Id: I75e01d3591d72faf288b093006d3801267afaa7c