init: refactor property triggers from a map to a simple vector
The only two uses of having an std::map for holding property triggers
were 1) to facilitate comparison of triggers when combining actions
and 2) to facilitate look up of previous triggers to ensure an action
only has one property trigger per property.  In all other cases, the
map is iterated over, so a simply std::vector would suffice.

However, 1) is no longer needed as we are no longer combining actions
and 2) is no longer useful as in the future, != property triggers will
require a more complex 'overlap' look up.

Test: Boot bullhead

Change-Id: I3a35042f7a99d6b973cc3b1fb025295f697360af