Update patch set 1
Patch Set 1:

I'm not thrilled about this change, but I'm not sure of any other particularly good options.  Update_verifier explicitly will want to loop infinitely, as it explicitly wants to prevent the system from booting any further (and writing to /data) if it is rebooting, so that behavior is expected.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>