sdcard: Use path arg in statfs operation.
FUSE passes the path used by the statfs syscall to the handler.
This patch uses it to figure out on which folder in the underlying mount
to call statfs. This allows cases in which statfs is called on a
subfolder of /sdcard (e.g. /sdcard/DCIM) to report the correct fs usage
in the case where this subfolder would have a different mount.

Test: boot bullhead, mount -t tmpfs -o size=1G tmpfs /data/media/0/DCIM,
Test: stat -f -c %f /sdcard/DCIM should report something different from
Test: stat -f -c %f /sdcard

Change-Id: I67dfbad507c207f439d51f992c7ddbefb7f331e4
Signed-off-by: roblabla <unfiltered@roblab.la>