logd: logd-reinit hand off /data/system/packages.h
logd starts before /data is mounted.  logd-reinit now also opens
/data/system/packages.h and hands the file descriptor off to logd for
processing along with the session key for decryption.

logd-reinit is started right after persistent properties are loaded
so that logd can reconfigure itself based on their values.

logd-reinit is started as one of the last members of late_start so
that a valid handoff can be made for a file descriptor with a valid
session key.

Test: gTest logd-unit-tests --gtest_filter=logd.statistics (success now)
Bug: 36645158
Bug: 37751120
Change-Id: I43ccb8b1cc8ed73bb9e6eb9385c9284c41258bda