ueventd: Break devices.cpp into discrete classes
devices.cpp handles too many things for creating one class.  This
change breaks it up into various files and classes.

Parsing is moved to ueventd_parser.cpp
Reading from the uevent socket is moved to a UeventListener class, in
its own file
ColdBooting is moved to it's own file
Firmware handling is moved to its own file
The remaining contents form a DeviceHandler class and the file is
renamed to device_handler.cpp

Test: boot bullhead, observe no differences in /dev and /sys
Test: init unit tests

Change-Id: I846a2e5995fbb344c7a8e349065c18a934fa6aba