ueventd: Break devices.cpp into discrete classes
devices.cpp handles too many things for creating one class.  This
change breaks it up into various files and classes.

Parsing is moved to ueventd_parser.cpp
Reading from the uevent socket is moved to a UeventListener class, in
its own file
ColdBooting is moved to a ColdBooter class in its own file
The remaining contents form a UeventHandler class and the file is
renamed to uevent_handler.cpp

Test: boot bullhead

Change-Id: I846a2e5995fbb344c7a8e349065c18a934fa6aba