init: fix first stage mount failure due to /dev/device-mapper not found
It has been reported that fs_mgr failed to open /dev/device-mapper
during the first stage mount. One suspicion is that other uevent was
reported as we didn't check whether the reported uevent is for device
mapper or not.

Fix this by checking the reported uevent->path matches that of device
mapper. Also print a log to test this hypothesis.

Bug: 37745254

Test: first stage mount /vendor with vboot 2.0 (avb) on bullhead in normal mode
Test: first stage mount /vendor with vboot 1.0 on sailfish in normal mode
Change-Id: I4a77093ec8f90a5ca981a088f34d082d0270533b