Adding 'uhid' permission for access to /dev/uhid.
This CL is in support of another CL c/2048848, topic
'Refactor hid command in /frameworks/base/cmds' in
internal master. Adding the permissions for
shell here to access uhid_node as part of the
new 'uhid' group.

Bug: 34052337
Test: Tested this change on aosp to make sure it builds. Could not
test on angler on aosp due to bluetooth not working (b/38135012).
Could not test on sailfish on aosp due to bootloop.

Change-Id: If9e100aa1262d689fb8adc5c0ce93f157c96399e