Update patch set 5
Patch Set 5:

Its a bug I guess. Its with the upstream kernel where uapi is selected. I believe it is where it is probably because no one wants to sign off on it never, or unlikely to, change or be affected by an architectural issue in the future. Should be made available at runtime via sysconf perhaps. Fixing that I considered outside the scope of this change as the value will affect many systems on Android (imagine it colliding with package UIDs)

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>