WIP: Support java trace dumps in tombstoned.
A rough (but fully) functional draft of the proposed changes to tombstoned.
It is now responsible for managing java ANR dumps in addition to native crashes.

The protocol is exactly the same, the type of request is inferred from the socket
it's received on. Having different sockets allows us to install different sets of
rules for each of those sockets; very few processes should need to write
or connect to the native crash socket.

TODO: Verify the provided PID with SO_PEERCRED

Bug: 32064548
Test: TBD

Change-Id: I1d62cc7a7035fd500c3e2b831704a2934d725e35