adb: discard incoming packets on the host until we hit CNXN.
When `adb kill-server` happens when a device is sending data to the
host, we can end up with a bunch of incoming data queued up that will be
read the next time we start adb, since we no longer reset the device
before connecting it. Work around this by reading and discarding data at
first until we hit an A_CNXN packet.

Bug: http://b/32952319
Test: python test_device.py
Change-Id: I1eed63bb782bbda3c1c5ac533442e7331dc41fb0