Update patch set 3
Patch Set 3:

> No this is a new patch.
 > 
 > Will wait for jhawkins@ for his preference. It would be easier to
 > follow the presubmit-hook that the entire project follows instead
 > of special ones for each sub-directory.

too late for that already. it's about half-and-half 4-space and 2-space. plus ones that are wrong like this, where it assumes 4 but is actually 2. (why did we add the default? to force everyone to make a decision one way or another, and at least stop the rot of randomly-formatted code.)

Patch-set: 3
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>