Update patch set 1
Patch Set 1:

If this is it's own tool, then it probably should be in a subdir like all of the other pieces.

I'm also not exactly sure the purpose of this tool is. It seems to be just printing the call stack of itself. The checkin message implies this can be used on any static binary, but doesn't appear to be the case.

Can you explain why you wrote this tool?

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>