RFC: debuggerd_client: Add a new initiate_dump method.
Debuggerd currently only supports a synchronous trigger_dump method
that sends a signal to a given process and waits for the dump to
complete. In the "new" stack dumping flow, we're considering just
sending the signal and not requiring the requester of the dump to
wait on the dump's completion. Tombstoned already keeps track of
time outs as well as the maximum number of concurrent dumps so there's
no need for the system_server to do so.

In very few cases (such as when collecting dropbox entries), we might
need to wait for traces to be completely written, but that can be
arranged separately in that code.

One wart / thing that needs discussion is this : debuggerd_trigger_dump
accepts a filename (always /data/anr/traces.txt, currently) and sets up
a tombstoned intercept to help redirect traces to that file. This method
will just use the default output paths, (/data/tombstones/tombstone_xx) so
we now have the unsavoury situation where :

(a) a file named tombstone_xx contains a native backtrace and not a tombstone
(b) this API doesn't really tell us where the tombstone is going -
callers will have to set up a separate inotify watch under the directory
to figure out which files were written. This already happens for
/data/anr.. files in both the old and new schemes so perhaps it's not a
huge deal.

Test: WIP
Change-Id: I0b2406524399b2fefd1cfb4f46621117897c7d32