add a performance benchmark for libziparchive
This patch adds two benchmarks that measure the performance of some
operations of libziparchive.

Both benchmarks are creating a temporary zip file containing file
names of uniformly distributed lengths.  The creation of the zip
file is not timed in the benchmarks.

- FindEntry_no_match tries to find an inexisting entry in the files
  of the zip archive, in order to force the code to examine all the
  files in the archive.

- Iterate_all_files uses the iterate function to list all the files
  in the archive.

To build the benchmark and to run on a device:

$ source build/envsetup.sh
$ lunch hikey-userdebug
$ cd system/core/libziparchive
$ mma &> build.log
$ cd -
$ adb push out/target/product/hikey/data/nativetest64/ziparchive-benchmarks/ziparchive-benchmarks /data/
$ adb shell /data/ziparchive-benchmarks
Run on (8 X 2.4 MHz CPU s)
1970-01-01 00:08:50
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
Benchmark                   Time           CPU Iterations
---------------------------------------------------------
FindEntry_no_match   12853318 ns   12853321 ns         54
Iterate_all_files    23822639 ns   23804554 ns         30

Change-Id: Ibdb524ba1c5ae55caddf0416ebbc09f8b6df0021