Update patch set 12
Patch Set 12:

Latest patchset adds proper checks that the subprocess have completed successfully.  

I had originally thought of having a sigchld handler and have a timeout and recovery if the subprocesses get stuck or crash, but that goes against a lot of the design here.  

For one, I would be simply using threads and not processes if it weren't for the setegid()/setfscreatecon() issues, so crashes in those threads would crash ueventd.  Secondly, that is actually the correct behavior, since ueventd is marked as 'critical' to init and we shouldn't hide crashes from it.

The timeout is similar, if a subprocess gets stuck, then something very bad must have happened and instead of adding logic to ueventd to try to handle it, it's better for us to hit the timeout that exists in init, which then reboots us to the bootloader.

That said this new code is still useful as it allows us to know if a subprocess has crashed or exited successfully.  With the new code, we'll only mark coldboot as successful if all subprocesses exited successfully.

I still need to test this a bit, but I wanted to upload for review anyway.

Patch-set: 12
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>