init: move killing of process groups to libprocessgroup
libprocessgroup kills the cgroup associated with a given pid and uid,
but not the POSIX process group associated with it.  This means that
to kill both, two of the same signals must be sent, which may cause
some issues.

This change has libprocessgroup kill the POSIX process group
associated with any pid that it encounters instead of just the pid
itself, and removes the kill for POSIX process groups from init.

Bug: 37853905
Test: Boot, kill zygote, reboot
Change-Id: Id1d96935745899b4c454c36c351ec16a0b1d3827