Update patch set 2
Patch Set 2:

(3 comments)

Higher level than all of this, depending on the timing requirements of vold, something like this could work:

on property:selinux.restorecon_recursive=*
  exec /system/bin/restorecon -R ${selinux.restorecon_recursive}
  setprop selinux.restorecon_recursive_finished ${selinux.restorecon_recursive}

And modify vold to query selinux.restorecon_recursive_finished.  It would be slower though, since the restorecon wouldn't be run until it made it through the action queue, so I'm not sure if it's a direction we want to take.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>