Update patch set 1
Patch Set 1:

Ok. Logan, what is your opinion on this? Should we mark this as vendor_avaiable or not? If you know any use of this from vendor lib, please let us know. 

If we keep marking this as vendor_available, we need to merge  https://android-review.googlesource.com/415499 and https://android-review.googlesource.com/415340

Patch-set: 1