Update patch set 8
Patch Set 8:

(6 comments)

I took some look at the property trigger solution and it's not going to be feasible.  Restorecon in vold is called from init_user0, which is called from an init builtin.  Unfortunately, that builtin already uses the exec mechanism of init, so we can't borrow it's logic and there will already be other items in the action queue, so we can't just queue it at the end and have it run in a reasonable amount of time.

Thanks for adding the queue; that fixes the issues I had in the previous patchsets.

There are some style comments that need to be addressed lastly.  If you run `git clang-format --style=file HEAD^` it will take care of most of them.  I wrote comments inline on the others.

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>