Update patch set 9
Patch Set 9:

As for the change to ReapOneProcess, apparently only services are ever expected to be launched and reaped by init.  So two options that come immediately to mind are to shoehorn this async process into a service or redesign the SIGCHLD handler.  The latter would probably involve keeping a map<pid_t,reaper_fn> as I had in my earlier patchsets.  That feels kinda overkill for this one-off, but I'd be willing to give it a whirl if you want.

Patch-set: 9
Reviewer: Gerrit User 1037023 <1037023@85c56323-6fa9-3386-8a01-6480fb634889>