Update patch set 9
Patch Set 9:

> I don't know what "git clang-format" is -- how do I acquire and run this?

I think all you'll need is:
https://github.com/llvm-mirror/clang/blob/master/tools/clang-format/git-clang-format

> Also, otapackage target seems to be gone in master.  How do you build an OTA zip?  Or is that gone now?

Hm it's working for me with aosp_bullhead-userdebug as my lunch target.  Perhaps it's something specific to the emulator that isn't working?

> So two options that come immediately to mind are to shoehorn this async process into a service or redesign the SIGCHLD handler. 

I'm not really saying as much that we should do anything different for this change, just saying that this isn't ideal for posterity.  I don't think shoe-horning this into a service helps us much, as services don't have a way to run commands once they've been terminated.

The two options that I see are 1) redesigning the sigchld handler, as you mentioned.  and I do want to do this for another reason anyway... but it's out of the scope of this change.
Or 2) we could keep this code within property_service.cpp if the subprocesses wrote a property back once they've completed.  They wouldn't want to use SetPropertyImpl for the reasons I mentioned before, but they should be able to write a property with the normal mechanisms, then property_service.cpp would catch it and act on it as a special case, similar to what we're already doing for selinux.restorecon_recursive.  The processes themselves would be reaped safely by service.cpp the same way orphaned processes are now.

I'm not actually sure if 2) would be better than what we have now though...

Patch-set: 9
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>