ueventd: remove PlatformDeviceList
In order to create symlinks for USB and block devices, the path for
their parent platform device must be known.

Previously, ueventd would save each platform device that it encounters
to a list and query this list when creating the symlinks.  That,
however, is racy particularly in the following scenario:

1) Kernel probes platform device
2) ueventd regenerates and processes uevents
3) Kernel probes block device
4) ueventd picks up the block device uevent during its uevent processing,
   without having regenerated the platform device, causing improper
   symlinks to be created.

This change stops storing the platform devices in a list, and instead
traverses up the directory structure for each USB or block device
until it reaches a platform device, defined as one whose subsystem is
the platform bus.  This fixes the race and simplifies the ueventd
code.

Bug: 62436493
Bug: 62681642
Test: Boot bullhead
Test: Unit tests are broken
Change-Id: I21636355d8e434f30e0cba568598a6cf139e67f9