[rfc] bootimg: better support for dtb
Currently the bootimg contains only information about kernel and
ramdisk, but not dtb. The lack of dtb information make it difficult
for bootloader to setup things properly, in a generic way.

There are several ways of providing/packaging dtb in Android environment.

1. Append dtb to the kernel image. It's up to the bootloader and kernel to
handle that. But the problem is that boot.img didn't provide explicit information
in the regard so that bootloader will know that the dtb is appended. In addition,
the ability of the bootloader and kernel to handle appended dtb
varies, say arm64 kernel image can't handle appended dtb (at the
moment).

2. Append dtb after the boot.img and use same partition as boot.img, or use a
dedicate dtb partition. Compare to 1, it is easier to extract/find the
dtb but the problem, still, is no explicit information available for
bootloader to know where the dtb is and handle it in a generic way.

There is a "second" area in the boot image, which is intended to put the
second stage bootloader, but in practices this seems were rarely used,
which is kind of waste :)

This patch propose a change to reuse the 2nd area to put the dtb. To achieve this,
we can reuse the "unused" field, which was designed for future expansion anyway! I wish
this is a problem/solution that can justify an extension. And, the "unused" field isn't
refrenced anywhere in master as of today (7/18/2016), so we can rename it as "flag".
We might also call it "dtb_in_second" to convey the intent more clearly,
but a general "flag" can leave room for future extension.

I have already have a PoC in the u-boot to use the second as the dtb and
boot an arm64 image, and I plan to upstream the uboot change but that
only make sense when this proposal is accepted.

[1]https://github.com/pierrchen/poplar-u-boot/commit/3099fa156d00929611802a64031a7513d6e1c752

v2:
To make things (e.g mkbootimg script) easier, and to encourage having seperate dtb, support
only one flag for dtb. When it is set to 1, means dtb is in the second area.

v3:
modify the mkbootimg script to support the flag.

v4:
1. mkbootimg: use os.path.basename instead of endswith.
2. update commit message.

Change-Id: Id8656009cb5a993bac91fe8c2b653bc96daa2c99
Signed-off-by: Bin Chen <bin.chen@linaro.org>