[rfc] bootimg: better support for dtb
Currently the bootimg contains only information about kernel and
ramdisk, but not dtb. The lack of dtb information make it difficult
for bootloader to setup things properly, in a generic way.

There are several ways of providing/packaging dtb in Android environment.

1. Appended to the kernel image, and thus resides in the boot.img. It's up
to the bootloader and kernel to handle that, but the problem is the boot.img
didn't provide anything information in the regard that the dtb is appended.
In addition, the ability of the bootloader and kernel to handle appended dtb
varies as well.

2.cat the dtb after the boot.img but use same partition, or use a
dedicate dtb partition. Compare to 1, it is easier to extract/find the
dtb but the problem, still, is no explicit information available for
bootloader to know where the dtb is and handle it in a generic way.

There is a second area in the boot image, which is intended to put the
second stage bootloader, seems this is rarely used. This patch propose a
change to reuse that area to put the dtb. To achieve this, I propose to
use the "unused" field, which was designed for future expansion! I wish
this is a problem/solution that can justfiy an extension. I also search the
"unused" field isn't used in current master as of today (7/18/2016), so we can
rename it as "flag".

I have already have a PoC in the u-boot to use the second as the dtb and
boot an arm64 image, and I plan to upstream the uboot change but that
only make sense when this proposal is accept as well. With those two, we
can have a upstream u-boot that works every android boot.img with dtb
support.

For other changes (build, mkbootimg script), I'd like to get some feedback
first before moving to far.

[1]https://github.com/pierrchen/poplar-u-boot/commit/3099fa156d00929611802a64031a7513d6e1c752

Change-Id: Id8656009cb5a993bac91fe8c2b653bc96daa2c99
Signed-off-by: Bin Chen <bin.chen@linaro.org>