libcutils_test: use set_cpuset_policy instead of set_sched_policy method
In some kernels, setting a new timerslack number does't lead to changing
of the cpusets. Instead of the implicit change by the set_sched_policy
method, this CL makes the get_sched_policy test use the
set_cpuset_policy method in order to change a mode of the scheduling
policy of the given process.

Bug: http://b/62598057, http://b/30675296
Test: libcutils_test passes.
Change-Id: Icd25cf99ce605bf908c517a638d01195dcea9181