sched_policy: change set_sched_policy to add support for cpusets
The set_sched_policy method also needs to add support the
cpusets-enabled kernel. In some kernels, setting a new timerslack number
does't lead to changing of the cpuset of the given process. Instead of
the implicit change by the set_sched_policy method, this CL makes the
set_sched_policy method utilizes the cpusets feature if it's available
to use in order to make sure to change the cpuset of the scheduling
policy of the given process.

Bug: http://b/62598057
Test: libcutils_test passes.
Change-Id: I9c33ba05caa786b21439326d82844b80cb8f8770