Update patch set 3
Patch Set 3:

Right, we have strongly suggested that the field should be set to zero, I'm not aware of a CTS test that would enforce this. If an OEM who is abusing this field is planning to update their devices to the next Android release, I don't feel it's unreasonable to ask them to deal with this change. After the previous change to the boot image header, it shouldn't come as a surprise to anyone that unused fields don't remain unused forever.

Patch-set: 3
Reviewer: Gerrit User 1058768 <1058768@85c56323-6fa9-3386-8a01-6480fb634889>