Update patch set 1
Patch Set 1:

I plan on sending changes to update those scripts.

However, I'm not planning on blocking the submission of this change on those changes. The reason being that it's unclear how / when / if those scripts are used, and for what purpose. This also means i won't be able to test them satisfactorily.

Patch-set: 1
Reviewer: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>