Update patch set 2
Patch Set 2:

(8 comments)

Generally looks good to me.  Thanks for writing this.  As I said earlier, I was hoping that we wouldn't need to do this, but it's pretty apparent now that we do.

I've got a few comments inline, the critical two are the one in action.cpp and the last one in reboot.cpp.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>