Update patch set 2
Patch Set 2:

(2 comments)

overall glad to see this, sorry I ran out of time to do anything on this. 

I'd like to see onshutdown <cmd> rather than adding fsck option to shutdown. I'm not sure how feasible that is though and it can always be in a seperate change later since this is all internal api. It's more work for sure, but it can start laying out the foundation for decoupling the hardcoded shutdown sequence.

Patch-set: 2
Reviewer: Gerrit User 1013433 <1013433@85c56323-6fa9-3386-8a01-6480fb634889>