bootstat.cpp: update time_since_last_boot to time_diff_last_boot
the vaule set for time_since_last_boot is the difference
between the value of last_boot_time_utc and current time,
and last_boot_time_utc is updated every boot at the boot
complete point.

So the value in fact is the difference between this boot
and last boot, so here update the variable name and event
name for source reading.

Test: manually with HiKey

Change-Id: I8991aabc9aaec7c6a52af0c577c4410fadb32ef1
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>